MATTER-4260 Fix 24 Coverity High Impact defects in chip-tool-web2
Address all 24 Coverity High Impact defects found in the webui-2_0 path
of the IMX_MATTER project. Changes span project code and vendored libs.

Project code (chip_tool_web.cpp, WebSocketClient.h):
- UNINIT: Initialize 12 Json::Value variables with explicit type
  (objectValue/arrayValue) to prevent uninitialized scalar reads
- RESOURCE_LEAK: Always call pclose(pipe) in generateOtaMessage
  regardless of killProcesses() return value

Crow library (json.h, multipart.h, mustache.h, query_string.h, routing.h):
- UNINIT: Zero-initialize rvalue members (start_, end_, lsize_, lremain_)
  in default and type constructors
- UNINIT: Value-initialize struct part in parse_section
- INTEGER_OVERFLOW: Use size_t loop counters with ptrdiff_t casts in
  pop_list() and pop_dict(); use size_t in Trie::add()
- USE_AFTER_MOVE: Remove redundant std::move before return in find();
  take std::string by value in rendered_template ctor

WebSocket++ (frame.hpp, endpoint.hpp):
- OVERRUN: Zero-initialize uint64_converter before partial copy in
  copy_payload(); add missing #include <cstring>
- RESOURCE_LEAK: Brace context.release() inside if(!ec) block in
  init_asio

CIDs: 53628846, 53628936, 53628940, 53628941, 53629067, 53629080,
53629088, 53629363, 53629381, 53629388, 53629501, 53629717, 53629765,
53629958, 53629989, 53630213, 53630224, 53630295, 53630379, 53630474,
53630609, 53630642, 53630775, 53630802

Reviewed-on: http://androidsource.nxp.com/project/32868

+++ examples/chip-tool/webui-2_0/MatterApplicationBackend/libs/crow/crow/query_string.h
@@ -420,14 +420,14 @@ namespace crow
             const size_t name_len = name.length();
             if (!ret.empty())
             {
-                for (unsigned int i = 0; i < key_value_pairs_.size(); i++)
+                for (size_t i = 0; i < key_value_pairs_.size(); i++)
                 {
                     std::string str_item(key_value_pairs_[i]);
                     if (str_item.find(name)==0) {
                       if (use_brackets && str_item.find("[]=",name_len)==name_len) {
-                        key_value_pairs_.erase(key_value_pairs_.begin() + i--);
+                        key_value_pairs_.erase(key_value_pairs_.begin() + static_cast<std::ptrdiff_t>(i--));
                       } else if (!use_brackets && str_item.find('=',name_len)==name_len ) {
-                           key_value_pairs_.erase(key_value_pairs_.begin() + i--);
+                           key_value_pairs_.erase(key_value_pairs_.begin() + static_cast<std::ptrdiff_t>(i--));
                        }
                     }
                 }
@@ -463,12 +463,12 @@ namespace crow
             std::unordered_map<std::string, std::string> ret = get_dict(name);
             if (!ret.empty())
             {
-                for (unsigned int i = 0; i < key_value_pairs_.size(); i++)
+                for (size_t i = 0; i < key_value_pairs_.size(); i++)
                 {
                     std::string str_item(key_value_pairs_[i]);
                     if (str_item.find(name_value)==0)
                     {
-                        key_value_pairs_.erase(key_value_pairs_.begin() + i--);
+                        key_value_pairs_.erase(key_value_pairs_.begin() + static_cast<std::ptrdiff_t>(i--));
                     }
                 }
             }
